requirements.txt
fastapi
uvicorn[standard]
sqlmodel
aiosqlite
async-exit-stack
python-multipart
aiofiles
passlib[bcrypt]
python-jose[cryptography]
httpx
litellm
crewai
langchain
langchain_huggingface
langchain_community
chromadb
PyPDF2
pandas
python-dotenv
pypdf